+++ lib/td_df_lib/format.ex
@@ -316,8 +316,13 @@ defmodule TdDfLib.Format do
 
   defp generate_default_return_map(value), do: %{"value" => value, "origin" => "default"}
 
+  def format_field(%{"content" => "", "type" => "url"}),
+    do: [%{"url_name" => "", "url_value" => ""}]
+
   def format_field(%{"content" => content, "type" => "url"}) do
-    [%{"url_name" => content, "url_value" => content}]
+    content
+    |> String.split("|", trim: true)
+    |> Enum.map(&format_url/1)
   end
 
   def format_field(%{
@@ -429,7 +434,7 @@ defmodule TdDfLib.Format do
 
         case nodes_found do
           [] ->
-            nil
+            {:error, :not_found}
 
           [%{"key" => key}] ->
             key
@@ -557,6 +562,27 @@ defmodule TdDfLib.Format do
   end
 
   defp format_hierarchy(hierarchy, _cardinality), do: hierarchy
+
+  defp format_url(content) do
+    case [
+      String.starts_with?(content, "["),
+      String.contains?(content, "] ("),
+      String.ends_with?(content, ")")
+    ] do
+      [true, true, true] ->
+        [url_name, url_value] =
+          content
+          |> String.replace_prefix("[", "")
+          |> String.replace_suffix(")", "")
+          |> String.split("] (")
+          |> Enum.map(&String.trim/1)
+
+        %{"url_name" => url_name, "url_value" => url_value}
+
+      _ ->
+        %{"url_name" => "", "url_value" => content}
+    end
+  end
 
   defp apply_cardinality(value = %{}, cardinality) when cardinality in ["*", "+"], do: [value]