+++ test/td_df_lib/format_test.exs
@@ -588,8 +588,67 @@ defmodule TdDfLib.FormatTest do
 
       assert formatted_value == [
                %{
-                 "url_name" => "https://google.es",
+                 "url_name" => "",
                  "url_value" => "https://google.es"
+               }
+             ]
+    end
+
+    test "format_field returns url with name wrapped" do
+      formatted_value =
+        Format.format_field(%{"content" => "[Google] (https://google.es)", "type" => "url"})
+
+      assert formatted_value == [
+               %{
+                 "url_name" => "Google",
+                 "url_value" => "https://google.es"
+               }
+             ]
+    end
+
+    test "format_field returns url with empty name" do
+      formatted_value =
+        Format.format_field(%{"content" => "[] (https://google.es)", "type" => "url"})
+
+      assert formatted_value == [
+               %{
+                 "url_name" => "",
+                 "url_value" => "https://google.es"
+               }
+             ]
+    end
+
+    test "format_field returns multiple url wrapped" do
+      formatted_value =
+        Format.format_field(%{
+          "content" => "[com] (www.com.com)|[] (www.net.net)|www.org.org",
+          "type" => "url"
+        })
+
+      assert formatted_value == [
+               %{
+                 "url_name" => "com",
+                 "url_value" => "www.com.com"
+               },
+               %{
+                 "url_name" => "",
+                 "url_value" => "www.net.net"
+               },
+               %{
+                 "url_name" => "",
+                 "url_value" => "www.org.org"
+               }
+             ]
+    end
+
+    test "format_field returns empty url for empty value" do
+      formatted_value =
+        Format.format_field(%{"content" => "", "type" => "url"})
+
+      assert formatted_value == [
+               %{
+                 "url_name" => "",
+                 "url_value" => ""
                }
              ]
     end
@@ -722,7 +781,11 @@ defmodule TdDfLib.FormatTest do
                Format.format_field(%{"content" => "foo", "type" => "user", "cardinality" => "+"})
 
       assert ["bar"] ==
-               Format.format_field(%{"content" => ["bar"], "type" => "user", "cardinality" => "+"})
+               Format.format_field(%{
+                 "content" => ["bar"],
+                 "type" => "user",
+                 "cardinality" => "+"
+               })
 
       assert "bar" ==
                Format.format_field(%{"content" => "bar", "type" => "user", "cardinality" => "1"})
@@ -833,15 +896,21 @@ defmodule TdDfLib.FormatTest do
     test "format_field with invalid content hierarchy  cardinality 1", %{
       hierarchy: %{id: id}
     } do
-      result =
-        Format.format_field(%{
-          "content" => "invalid",
-          "type" => "hierarchy",
-          "cardinality" => "1",
-          "values" => %{"hierarchy" => %{"id" => id}}
-        })
+      assert {:error, :not_found} =
+               Format.format_field(%{
+                 "content" => "invalid",
+                 "type" => "hierarchy",
+                 "cardinality" => "1",
+                 "values" => %{"hierarchy" => %{"id" => id}}
+               })
 
-      assert is_nil(result)
+      assert "" =
+               Format.format_field(%{
+                 "content" => "",
+                 "type" => "hierarchy",
+                 "cardinality" => "1",
+                 "values" => %{"hierarchy" => %{"id" => id}}
+               })
     end
 
     test "format_field with content hierarchy multiple node with cardinality *", %{