+++ CHANGELOG.md
@@ -6,6 +6,18 @@
 
 - [TD-5119] add validation for template dependant fields
 
+## [7.1.1] 2025-02-03
+
+### Added
+
+- [TD-6022] add validation for template hierarchy fields
+
+## [7.1.0] 2025-01-29
+
+### Changed
+
+- [TD-6921] Change URL type field parser
+
 ## [7.0.3] 2025-01-13
 
 ### Changed